requirements.txt
loggerric
socketric
pillow
pyperclip